+++ modules/task_3/kulandin_d_quick_sort/quick_sort.h
@@ -4,8 +4,8 @@
 #include <vector>
 
 int* randomVector(int sz);
-int* copyVector(int* a, size_t sz);
-void division(int* a, int l, int r, size_t* t);
-void parallelQuickSort(int* a, size_t length);
+int* copyVector(int* a, int sz);
+void division(int* a, int l, int r, int* t);
+void parallelQuickSort(int* a, int length);
 
 #endif  // MODULES_TASK_3_KULANDIN_D_QUICK_SORT_QUICK_SORT_H_
